// src/services/models/userResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Swipe Pad API
 * OpenAPI spec version: 0.0.12
 */

export interface UserResponse {
  readonly id: number;
  readonly created_at: string;
  readonly updated_at: string;
  readonly walletAddress: string;
  readonly name: string;
  readonly nullifierHash: string;
  readonly image: string;
  readonly bio: string;
  readonly totalFollowers: number;
  readonly totalFollowing: number;
}